backend/requirements.txt
 django == 3.0.8  
djangorestframework
mysqlclient >= 1.4.3